moved API interfaces to interfaces folder in api/
