+++ vmodules/api/get_guild_config.ts
@@ -1,5 +1,5 @@
 import memory from '../core/api_memory';
-import APIHandler from '../util/interfaces/APIHandler';
+import APIHandler from './interfaces/APIHandler';
 
 const handler: APIHandler = {
   route: `/guildcfg`,
