+++ vmodules/api/interfaces/APIMessage.ts
@@ -1,0 +1,3 @@
+export default interface APIMessage {
+  doc?: string | null
+}
